yolo_linear/requirements.txt
torch
torchvision
matplotlib
opencv-contrib-python
Pillow
numpy
tqdm
mtcnn